refactor(server): split ws_session.py into session/handlers/runner
ws_session.py owned per-connection state, RPC handlers, and the
receiver / sender / generator workers all in one ~1k-line file.
Split along the natural seams:

- ws_session.py: Connection, ControlState, build_error_message
- ws_handlers.py: handle_init, handle_check_seed_safety,
  load_seed_from_data, build_init_response_data
- ws_runner.py: run_receiver, run_sender, run_generator,
  reset_engine, _PendingFlush

Each module now has one area of responsibility. The lazy
`save_safety_cache` import (kept out of module-load to avoid
dragging torch in transitively) moves with the handlers.
