+++ server-components/ws_handlers.py
@@ -1,0 +1,285 @@
+"""
+Typed message handlers for the WebSocket protocol.
+
+Each handler takes the values it needs explicitly (Connection +
+engine refs) and returns / produces typed Pydantic models. Stateless
+dispatchers — they read/write `conn.X` and call into engine_manager
+via the explicitly-passed handles, but own no per-handler state.
+
+`run_receiver` (in ws_runner.py) dispatches inbound messages to these;
+the pre-init handshake in server.py also calls them directly.
+
+This module is strict-typed by construction — none of the legacy ignore
+rules in pyproject.toml fire on this code. Keep it that way.
+"""
+
+import asyncio
+import base64
+import hashlib
+import io
+import logging
+import time
+from typing import TYPE_CHECKING
+
+from PIL import Image
+
+from action_logger import ActionLogger
+from app_state import AppState, SafetyCacheEntry
+from protocol import (
+    CheckSeedSafetyRequest,
+    CheckSeedSafetyResponseData,
+    InitRequest,
+    InitResponseData,
+    MessageId,
+    RpcError,
+    RpcSuccess,
+    StageId,
+    SystemInfo,
+    rpc_err,
+    rpc_ok,
+)
+from ws_session import Connection
+
+if TYPE_CHECKING:
+    from engine_manager import WorldEngineManager
+    from safety import SafetyChecker
+
+logger = logging.getLogger(__name__)
+
+
+def _compute_bytes_hash(data: bytes) -> str:
+    return hashlib.sha256(data).hexdigest()
+
+
+def build_init_response_data(world_engine: "WorldEngineManager", system_info: SystemInfo) -> InitResponseData:
+    """Pack post-warmup session metrics into the typed init RPC response."""
+    return InitResponseData(
+        model=world_engine.model_uri or "",
+        inference_fps=world_engine.inference_fps,
+        system_info=system_info,
+    )
+
+
+async def handle_check_seed_safety(
+    state: AppState,
+    req: CheckSeedSafetyRequest,
+) -> RpcSuccess[CheckSeedSafetyResponseData] | RpcError:
+    """Check whether a seed image passes the NSFW classifier. Caches by
+    SHA-256 of the raw bytes; results survive across server restarts via
+    the on-disk safety cache."""
+    if not req.image_data:
+        return rpc_err(req.req_id, error=MessageId.SEED_MISSING_DATA.value)
+
+    try:
+        image_bytes = base64.b64decode(req.image_data)
+    except Exception as e:
+        return rpc_err(req.req_id, error=f"Invalid base64 data: {e}")
+
+    img_hash = _compute_bytes_hash(image_bytes)
+
+    cache = state.safety_hash_cache
+    if img_hash in cache:
+        cached = cache[img_hash]
+        return rpc_ok(
+            req.req_id,
+            CheckSeedSafetyResponseData(is_safe=cached.is_safe, hash=img_hash),
+        )
+
+    safety_checker = state.safety_checker
+    try:
+
+        def _check_safety():
+            pil_img = Image.open(io.BytesIO(image_bytes)).convert("RGB")
+            return safety_checker.check_pil_image(pil_img)
+
+        safety_result = await asyncio.to_thread(_check_safety)
+    except Exception as e:
+        logger.error(f"Safety check failed: {e}")
+        return rpc_err(req.req_id, error=f"Safety check failed: {e}")
+
+    # Function-level import: `safety` pulls torch transitively, so we don't
+    # want to drag it into the module graph at import-time. The heavy
+    # try-block in server.py is the only place that should trigger the
+    # torch / transformers import waterfall.
+    from safety import save_safety_cache
+
+    is_safe = safety_result.get("is_safe", False)
+    cache[img_hash] = SafetyCacheEntry(
+        is_safe=is_safe,
+        scores=safety_result.get("scores", {}),
+        checked_at=time.time(),
+    )
+    save_safety_cache(cache)
+
+    return rpc_ok(req.req_id, CheckSeedSafetyResponseData(is_safe=is_safe, hash=img_hash))
+
+
+async def load_seed_from_data(
+    conn: Connection,
+    world_engine: "WorldEngineManager",
+    safety_checker: "SafetyChecker",
+    image_data_b64: str | None,
+    seed_filename: str | None = None,
+) -> bool:
+    """Validate safety and load a seed from base64 image data.
+
+    Returns True iff the seed was loaded (or already loaded and matched).
+    Failure paths surface a typed warning over the websocket and return False.
+    """
+    if not image_data_b64:
+        logger.warning(f"[{conn.client_host}] Missing seed image data")
+        await conn.send_warning(MessageId.SEED_MISSING_DATA)
+        return False
+
+    try:
+        image_bytes = base64.b64decode(image_data_b64)
+    except Exception as e:
+        logger.warning(f"[{conn.client_host}] Invalid base64 seed data: {e}")
+        await conn.send_warning(MessageId.SEED_INVALID_DATA)
+        return False
+
+    img_hash = _compute_bytes_hash(image_bytes)
+
+    # Same seed already loaded? Skip the safety inference + load entirely.
+    if img_hash == conn.current_seed_hash:
+        logger.info(f"[{conn.client_host}] Seed unchanged (hash match), skipping reload")
+        return True
+
+    # Safety check (cache hit, then live inference if missing)
+    cache = conn.state.safety_hash_cache
+    if img_hash in cache:
+        cached = cache[img_hash]
+        if not cached.is_safe:
+            logger.warning(f"[{conn.client_host}] Seed marked as unsafe (cached)")
+            await conn.send_warning(MessageId.SEED_UNSAFE)
+            return False
+    else:
+        try:
+
+            def _check_safety():
+                pil_img = Image.open(io.BytesIO(image_bytes)).convert("RGB")
+                return safety_checker.check_pil_image(pil_img)
+
+            safety_result = await asyncio.to_thread(_check_safety)
+        except Exception as e:
+            logger.warning(f"[{conn.client_host}] Safety check failed: {e}")
+            await conn.send_warning(MessageId.SEED_SAFETY_CHECK_FAILED)
+            return False
+
+        # Function-level import: see comment in handle_check_seed_safety.
+        from safety import save_safety_cache
+
+        is_safe = safety_result.get("is_safe", False)
+        cache[img_hash] = SafetyCacheEntry(
+            is_safe=is_safe,
+            scores=safety_result.get("scores", {}),
+            checked_at=time.time(),
+        )
+        save_safety_cache(cache)
+
+        if not is_safe:
+            logger.warning(f"[{conn.client_host}] Seed marked as unsafe")
+            await conn.send_warning(MessageId.SEED_UNSAFE)
+            return False
+
+    # Load the seed onto the engine
+    display_name = seed_filename or img_hash[:12]
+    logger.info(f"[{conn.client_host}] Loading seed '{display_name}'")
+    loaded_frame = await world_engine.load_seed_from_base64(image_data_b64)
+    if loaded_frame is None:
+        logger.error(f"[{conn.client_host}] Failed to load seed")
+        await conn.send_warning(MessageId.SEED_LOAD_FAILED)
+        return False
+
+    world_engine.seed_frame = loaded_frame
+    world_engine.original_seed_frame = loaded_frame
+    conn.current_seed_hash = img_hash
+    conn.current_seed_filename = seed_filename
+    logger.info(f"[{conn.client_host}] Seed loaded successfully")
+    return True
+
+
+async def handle_init(
+    conn: Connection,
+    world_engine: "WorldEngineManager",
+    safety_checker: "SafetyChecker",
+    req: InitRequest,
+    *,
+    is_game_loop: bool = False,
+) -> tuple[bool, bool]:
+    """Apply an InitRequest's deltas to the connection / engine.
+
+    Returns `(ready, seed_loaded)`: `ready` means the session has a
+    seed frame and can begin generating; `seed_loaded` means a fresh
+    seed was applied in this call.
+
+    `model_fields_set` distinguishes "field absent" from "field present
+    and explicitly None", so partial updates (`{"action_logging": false}`
+    while leaving everything else untouched) work as the renderer expects.
+    """
+    present = req.model_fields_set
+    model_uri = (req.model or "").strip()
+    seed_data = req.seed_image_data
+    seed_filename = req.seed_filename
+    quant = req.quant
+
+    # Update flags
+    if "scene_authoring" in present and req.scene_authoring is not None:
+        conn.scene_authoring_requested = req.scene_authoring
+    if "action_logging" in present and req.action_logging is not None:
+        conn.action_logging_requested = req.action_logging
+    if "video_recording" in present and req.video_recording is not None:
+        conn.video_recording_requested = req.video_recording
+    if "video_output_dir" in present:
+        conn.video_output_dir = req.video_output_dir
+    if "biome_version" in present:
+        conn.biome_version = req.biome_version
+    if "cap_inference_fps" in present and req.cap_inference_fps is not None:
+        conn.cap_inference_fps = req.cap_inference_fps
+
+    # Sync recorder lifecycle with requested state during gameplay
+    if is_game_loop:
+        if conn.action_logging_requested and conn.action_logger is None:
+            conn.action_logger = ActionLogger(conn.client_host)
+            conn.start_action_log_segment(world_engine)
+            logger.info(f"[{conn.client_host}] Action logging enabled")
+        elif not conn.action_logging_requested and conn.action_logger is not None:
+            conn.action_logger.end_segment()
+            conn.action_logger = None
+            logger.info(f"[{conn.client_host}] Action logging disabled")
+
+        if conn.video_recording_requested and conn.video_recorder is None:
+            conn.start_video_segment(world_engine)
+            logger.info(f"[{conn.client_host}] Video recording enabled")
+        elif not conn.video_recording_requested and conn.video_recorder is not None:
+            conn.video_recorder.end_segment()
+            conn.video_recorder = None
+            logger.info(f"[{conn.client_host}] Video recording disabled")
+
+    # Model delta — reload if model URI or quantization changed.
+    # The engine must be loaded before the seed so that seed_target_size
+    # and temporal_compression are resolved from the actual model config.
+    model_changed = False
+    quant_changed = "quant" in present and quant != world_engine.quant
+    if model_uri and (model_uri != world_engine.model_uri or quant_changed):
+        verb = "Live model switch" if is_game_loop else "Requested model"
+        logger.info(f"[{conn.client_host}] {verb}: {model_uri} (quant={quant})")
+        world_engine.set_progress_callback(conn.push_progress, conn.main_loop)
+        await world_engine.load_engine(model_uri, quant=quant)
+        world_engine.set_progress_callback(None)
+        world_engine.seed_frame = None
+        conn.perceptual_frame_count = 0
+        conn.max_perceptual_frames = (world_engine.n_frames - 2) * world_engine.temporal_compression
+        model_changed = True
+        logger.info(f"[{conn.client_host}] Model loaded: {world_engine.model_uri}")
+
+    # Seed delta
+    seed_loaded = False
+    if seed_data:
+        seed_loaded = await load_seed_from_data(conn, world_engine, safety_checker, seed_data, seed_filename)
+
+    if model_changed and not seed_loaded and not world_engine.seed_frame:
+        await conn.send_stage(StageId.SESSION_WAITING_FOR_SEED)
+
+    ready = seed_loaded or (world_engine.seed_frame is not None)
+    return ready, seed_loaded
